# perf: replace encoding/json with goccy/go-json