decoding the bulk response is taking ~9% of the total time in local testing.
Replacing the stdlib json decoder with a custom library seems to have some performance benefits:

```
                                      │  before.txt   │              after.txt               │
                                      │    sec/op     │    sec/op     vs base                │
Appender/NoCompression-20                478.2n ±  3%   460.6n ±  5%        ~ (p=0.052 n=10)
Appender/NoCompressionScaling-20         872.6n ± 12%   819.1n ± 19%        ~ (p=0.529 n=10)
Appender/BestSpeed-20                   1366.5n ±  7%   758.5n ±  3%  -44.49% (p=0.000 n=10)
Appender/BestSpeedScaling-20            1311.5n ±  6%   721.8n ±  4%  -44.96% (p=0.000 n=10)
Appender/DefaultCompression-20           1.770µ ±  1%   1.376µ ±  2%  -22.26% (p=0.000 n=10)
Appender/DefaultCompressionScaling-20    1.763µ ±  2%   1.386µ ±  4%  -21.38% (p=0.000 n=10)
Appender/BestCompression-20              1.788µ ±  2%   1.368µ ±  5%  -23.49% (p=0.000 n=10)
Appender/BestCompressionScaling-20       1.778µ ±  2%   1.382µ ±  3%  -22.30% (p=0.000 n=10)
geomean                                  1.285µ         963.8n        -24.98%

                                      │  before.txt   │               after.txt                │
                                      │      B/s      │      B/s        vs base                │
Appender/NoCompression-20               267.2Mi ±  3%    277.5Mi ±  5%        ~ (p=0.052 n=10)
Appender/NoCompressionScaling-20        146.5Mi ± 11%    156.0Mi ± 16%        ~ (p=0.529 n=10)
Appender/BestSpeed-20                   93.53Mi ±  7%   168.48Mi ±  3%  +80.14% (p=0.000 n=10)
Appender/BestSpeedScaling-20            97.45Mi ±  6%   177.04Mi ±  4%  +81.68% (p=0.000 n=10)
Appender/DefaultCompression-20          72.20Mi ±  1%    92.87Mi ±  3%  +28.64% (p=0.000 n=10)
Appender/DefaultCompressionScaling-20   72.49Mi ±  2%    92.21Mi ±  4%  +27.21% (p=0.000 n=10)
Appender/BestCompression-20             71.48Mi ±  2%    93.42Mi ±  4%  +30.69% (p=0.000 n=10)
Appender/BestCompressionScaling-20      71.88Mi ±  2%    92.50Mi ±  3%  +28.69% (p=0.000 n=10)
geomean                                 99.49Mi          132.6Mi        +33.28%

                                      │ before.txt │             after.txt              │
                                      │    B/op    │    B/op     vs base                │
Appender/NoCompression-20               513.5 ± 0%   872.0 ± 0%  +69.81% (p=0.000 n=10)
Appender/NoCompressionScaling-20        532.0 ± 1%   880.5 ± 0%  +65.51% (p=0.000 n=10)
Appender/BestSpeed-20                   984.0 ± 0%   899.0 ± 1%   -8.64% (p=0.000 n=10)
Appender/BestSpeedScaling-20            984.0 ± 0%   899.0 ± 1%   -8.64% (p=0.000 n=10)
Appender/DefaultCompression-20          981.5 ± 1%   883.0 ± 1%  -10.04% (p=0.000 n=10)
Appender/DefaultCompressionScaling-20   977.0 ± 5%   872.0 ± 2%  -10.75% (p=0.000 n=10)
Appender/BestCompression-20             982.5 ± 6%   884.0 ± 1%  -10.03% (p=0.000 n=10)
Appender/BestCompressionScaling-20      977.5 ± 5%   884.0 ± 1%   -9.57% (p=0.000 n=10)
geomean                                 838.2        884.1        +5.48%

                                      │ before.txt  │              after.txt               │
                                      │  allocs/op  │ allocs/op   vs base                  │
Appender/NoCompression-20                7.000 ± 0%   7.000 ± 0%        ~ (p=1.000 n=10) ¹
Appender/NoCompressionScaling-20         7.000 ± 0%   7.000 ± 0%        ~ (p=1.000 n=10) ¹
Appender/BestSpeed-20                   10.000 ± 0%   7.000 ± 0%  -30.00% (p=0.000 n=10)
Appender/BestSpeedScaling-20            10.000 ± 0%   7.000 ± 0%  -30.00% (p=0.000 n=10)
Appender/DefaultCompression-20          10.000 ± 0%   7.000 ± 0%  -30.00% (p=0.000 n=10)
Appender/DefaultCompressionScaling-20   10.000 ± 0%   7.000 ± 0%  -30.00% (p=0.000 n=10)
Appender/BestCompression-20             10.000 ± 0%   7.000 ± 0%  -30.00% (p=0.000 n=10)
Appender/BestCompressionScaling-20      10.000 ± 0%   7.000 ± 0%  -30.00% (p=0.000 n=10)
geomean                                  9.147        7.000       -23.47%
```